Added styling to character boxes and put CharacterBox in its own file
