+++ react-sorting-hat/src/App.css
@@ -36,11 +36,27 @@
 
 .cardsBox {
   display: flex;
-  flex-flow: column wrap;
+  flex-flow: row wrap;
   justify-content: center;
-  align-items: center;
+  align-items: stretch;
 }
 .characterBox {
-  border: 1px solid black;
   margin: 1rem;
+  padding: 1rem;
+  width: 33%;
+  border-radius: 2px;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
+  display: flex;
+  flex-flow: column nowrap;
+
+}
+
+.characterBox h3 {
+  color: #FFFFFF;
+  text-shadow: 0 0 10px #FFFFFF;
+  letter-spacing: 0.45rem;
+}
+
+.characterBox img {
+  border: 2px solid black;
 }
