+++ react-sorting-hat/src/App.js
@@ -10,7 +10,7 @@ class App extends Component {
   constructor() {
     super();
     this.state = {
-      house: 'G'
+      house: 'R'
     }
   }
