+++ react-sorting-hat/src/houseData.js
@@ -26,7 +26,7 @@ const houseData = [
         ghost: 'the Grey Lady, the daughter of Rowena Ravenclaw',
         commonRoomLocation: 'in a tower on the west side of the castle.',
         commonRoomDescription: 'a wide, airy, circular room. Graceful arched windows punctuated the walls, which were hung with blue- and-bronze silks. The ceiling was domed and painted with stars, which were echoed in the midnight-blue carpet. There were tables, chairs, and bookcases, and in a niche opposite the door stood a tall statue of white marble.',
-        commonRoomEntrance: 'Ravenclaw students must answer a riddle as opposed to giving a password to enter their dormitories. This riddle, however, can be answered by non-Ravenclaws.',
+        commonRoomEntrance: 'entered by answering a riddle, as opposed to giving a password. This riddle, however, can be answered by non-Ravenclaws.',
         headmaster: 'Filius Flitwick',
         peopleType: 'Ravenclaws value intelligence, knowledge, and wit.',
         animal: 'eagle',
@@ -41,7 +41,7 @@ const houseData = [
         ghost: 'The Fat Friar',
         commonRoomLocation: 'somewhere in the basement, near the castle\'s kitchens.',
         commonRoomDescription: 'a cosy and welcoming place with yellow hangings, fat armchairs, and underground tunnels that lead to the dormitories, which have perfectly round doors, similar to barrel tops.',
-        commonRoomEntrance: 'It can be accessed by tapping the barrel two from the bottom, middle of the second row in the rhythm of "Helga Hufflepuff".',
+        commonRoomEntrance: 'can be accessed by tapping the barrel two from the bottom, middle of the second row in the rhythm of "Helga Hufflepuff".',
         headmaster: 'Pomona Sprout',
         peopleType: 'Hufflepuffs value hard work, dedication, patience, loyalty, and fair play.',
         animal: 'badger',
