+++ react-sorting-hat/src/index.css
@@ -9,6 +9,8 @@ body {
 h1 {
   font-family: 'Ruslan Display', cursive;
   font-size: 2.5rem;
+  letter-spacing: 0.2rem;
+  word-spacing: 1rem;
 }
 
 h2 {
@@ -21,7 +23,8 @@ h3 {
 }
 
 p { 
-  font-size: 1rem;
+  font-size: 1.2rem;
+  line-height: 2rem;
 }
 
 a {
@@ -68,5 +71,6 @@ button {
 
 img {
   max-width: 100%;
+  border-radius: 2px;
 }
